move 'FeedImageMapper' to its own file
