+++ FeedAPIChallenge/Feed API/RemoteFeedLoader.swift
@@ -30,32 +30,3 @@ public final class RemoteFeedLoader: FeedLoader {
 		}
 	}
 }
-
-private class FeedImageMapper{
-	private struct root: Decodable{
-		let items: [Image]
-		
-		var feedItems: [FeedImage]{
-			items.map {$0.feedImage}
-		}
-	}
-	
-	private struct Image: Decodable {
-		let image_id: UUID
-		let image_desc: String?
-		let image_loc: String?
-		let image_url: URL
-		
-		var feedImage: FeedImage {
-			return FeedImage(id: image_id, description: image_desc, location: image_loc, url: image_url)
-		}
-	}
-	
-	static fileprivate func map( _ data: Data, _ response: HTTPURLResponse) -> FeedLoader.Result {
-		if response.statusCode == 200, let root = try? JSONDecoder().decode(root.self, from: data){
-			return .success(root.feedItems)
-		}else{
-			return .failure(RemoteFeedLoader.Error.invalidData)
-		}
-	}
-}
